Update styling of Featured Fighters page
